メモの新規作成機能をDbOperatorクラスのメソッドとして切り出した
・DBへのアクセスはすべてDbOperatorの責務としたいため切り出すこととした
