+++ 04.class/memo_app.js
@@ -1,15 +1,8 @@
-const Memo = require('./memo.js');
-const ReferenceEnquirer = require('./reference_enquirer');
+const DbOperator = require('./db_operator.js');
+const dbOperator = new DbOperator();
 const ListProvider = require('./list_provider');
+const ReferenceEnquirer = require('./reference_enquirer');
 const DestroyEnquirer = require('./destroy_enquirer');
-const readline = require('node:readline/promises');
-const sqlite3 = require('sqlite3').verbose();
-const db = new sqlite3.Database('memo_app');
-//const { Select } = require('enquirer');
-const reader = readline.createInterface({
-  input: process.stdin,
-  output: process.stdout
-});
 
 let selectedOption = ''
 
@@ -28,18 +21,5 @@ switch (process.argv[2]) {
 if (process.stdin.isTTY) {
   selectedOption.show();
 } else {
-  // 入力の受け取り
-  let lines = [];
-  reader.on('line', (line) => {
-    lines.push(line);
-  });
-  
-  // Memoインスタンスの生成と保存
-  reader.on("close", () => {
-    const memo = new Memo(lines);
-    db.serialize(() => {
-      db.run('CREATE TABLE if not exists memos(id INTEGER PRIMARY KEY, body TEXT)');
-      db.run('INSERT INTO memos (body) values(?)', [memo.lines]);
-    });
-  });
+  dbOperator.create();
 };
